requirements.txt
Django==1.10
djangorestframework==3.4.6
Pillow==3.3.1
pkg-resources==0.0.0
